* src/attack.c (launch_plugin): Ignore script_mandatory_keys requirements during network-wide scanning phase.
-- Additional information:
This change allow us to make use of script_mandatory_keys for the hosts scan
phase (and therefore heavily reduce the load), without having it preventing
scripts from running during the first phase.

Keys are usually required only for the second scan phase. If it's not the case
that will only be one extra process spawn per network, instead of one per
target. The extra processes can then look for the required key and directly
exit if not found.


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-scanner@12513 423fd1db-d629-0410-8442-d21db03e70f4