Remove any use of the "OID" arglist element of a plugin arglist structure.
* src/attack.c (launch_plugin): The plugins' arglist name is
the OID, so use this instead of aglist element "OID".

* src/pluginlaunch.c (process_internal_msg, update_running_processes,
read_running_processes, plugin_launch): The plugins' arglist name is
the OID, so use this instead of aglist element "OID".

* src/comm.c (send_plug_info, plugin_send_infos, qsort_cmp,
_get_plug_by_oid): The plugins' arglist name is
the OID, so use this instead of aglist element "OID".

* src/pluginscheduler.c (hash_add, enable_plugin_and_dependencies,
plugins_scheduler_init): The plugins' arglist name is the OID, so
use this instead of aglist element "OID".



git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-scanner@20864 423fd1db-d629-0410-8442-d21db03e70f4