openvas-nasl-lint does not go deeper through a declared function if it is not called/used. If a not used function has unsatisfied dependencies, no errors will show.
* nasl/lint.c (nasl_lint_def): Receive a new argument. Add a comment.
Check if a declared function will be used, before to continue to going
deeper in the tree.

* nasl/lint.c (nasl_lint_call): Receive a new argument. Check if a
declared function will be used, before to continue to going deeper in
the tree.

* nasl/lint.c (nasl_lint_defvar): Receive a new argument. Check if a
declared function will be used, before to continue to going deeper in
the tree.

* nasl/lint.c (make_call_func_list): New function. It adds to a list
all called functions.

* nasl/lint.c (nasl_lint): New GSList. Call make_call_func_list().
Pass as argument the new GSList to functions. Free the new list.


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-scanner@29610 423fd1db-d629-0410-8442-d21db03e70f4