Check the version of libgcrypt
Previously CMake only checked for the presence of libgcrypt, causing
build failures when the installed libgcrypt was too old and did not
provide the required symbols.

This commit adds a version check for libgcrypt which will fail if the
version is below 1.6. This version requirement is now documented in the
INSTALL file as well.

This commit was adapted from greenbone/gvm-libs#41 (which in turn was in
reaction to greenbone/gvm-libs#40) since the relevant checks have moved
from gvm-libs to openvas-scanner in the master branch.