Move check for pkg-config to where it is used
As the top-level CMakeLists.txt no longer checks dependencies directly
it makes no sense for it to consider a missing pkg-config binary a fatal
error.

Instead, it is the job of the second-level CMakeLists.txt files which
are actually using pkg-config to ensure it is available and fail if not.

This change makes openvas-scanner consistent with gvm-libs. It also
means pkg-config is no longer a (useless) dependency for generating
source code documentation.