Rename setting "kb_location" to "db_address"
Meanhile the redis database is not only used to store temporarily
some information about a host during a scan.
It is used as a general database backend, for example to store the
NVT meta data.

This name change reflects the new usage. Also it is more of an "address"
rather than a "location".