Fix parameter pollution in certain nasl functions.
Some NASL functions use get_*_var_by_name() instead of
get_*_local_var_by_name(), which "climbs" in the context when searching
for a variable ie. if a parameter 'foo' is not provided, the function
would use the variable 'foo' if found.

Unify the behaviour by replacing all get_*_local_var_by_name()
with get_*_var_by_name() calls, and change the later to not climb
in the nasl context when searching for the variable.