filter by port
Instead of filtering for every host individually we filter by port.
This way we can filter for many hosts without running in memory issues
with too big filter descriptions. A new hashtable of the target hosts
is used for not including alive hosts we do not want to scan by mistake.