Prevent deadlock, ensure that no locks are being held at the time of forking.
During testing, the child deadlock was reproduce, directly after fork, when logging the message
"Testing xxx.xxx.xxx.xxx (Vhost...."

Since alive detection is running in a thread, it can lock a mutex when the main process is forking a new child.
This mutex is copied into the child process in its locked state, and any child that tries to lock the mutex waits forever.
This patch prevent the child deadlock.