Use internal function name in debug messages too
nasl_get_function_name only return the function
of the nasl script in which the current function
is called.
Now we also include __fun__ in the debug message.
When the internal function is not called from
inside a nasl function the placeholder
"script_main_function" is used as nasl function
name.