Add function for generating ipv6 mask from prefix
Original ipv6addrmask() has several limitations
and bugs. It only had a byte size resolution.
E.g. masking with /124 yielded an erroneous result.
It included some hacky way of clearing an uint32_t
which seems to be not supported with modern compilers
anymore which resulted in not clearing anything if
the prefix was a multiple of 32 on some compilation
configurations.