Move technique switch outside of loop
procroutetechnique is the default technique for
getting the interface we want to use for some given
destination addr.
This technique is represented as a static enum.
Previously the technique was set to fallback
method connectsockettechnique even though an
appropriate interface was found because
the fallback method was activated after the first
route in /proc/net/ipv6_route was checked and not
after all routes were checked. Every additional
invocation of the function therefore used the
fallback method as new default.