Changed defaults for installation locations
* SYSCONFDIR is /etc by default now
* LOCALSTATEDIR is /var by default now
* OPENVAS_RUN_DIR is /run/ospd by default now
* OPENVAS_FEED_LOOK_PATH is /var/lib/openvas/feed-update.lock by default
  now

/run/ospd is used because normally openvas is started from ospd-openvas
which by default uses the /run/ospd path in the systemd service file.
This ensures that the directory for the OPENVAS_RUN_DIR exists at
runtime.

/var/lib/openvas/feed-update.lock is used because the sync scripts may
be started when the services are stopped an /run/ospd (or an other
directory in /run) is not available anymore.

(cherry picked from commit adc9fecb494d9a84c62ee506c8c9b90b01e1aa54)