Check if the max running process was reached or even exceed.
Fix SC-177

On some cases, the parallel scan is disabled and,
the max_running_processes variable is set to 1, while the current number of
running process is grether than 1. Therefore, condition in the while loop are
not satisfied and never wait for free slot. At the end, the process table is full,
and this produced the host scan process to stop, because not free slot for a new process.