Change: Instead of using g_memdup2 set deprecation for it as warning
Since g_memdup2 will use a 32bit number instead of a 64bit there could
be the case that it will allocate less memory than anticipated could due
to lack of control end up in an overflow scenario.

To prevent that we will leave it as g_memdup but redurce the severity
down to a warning instead of an error.