Fix plugins upload.
When a plugin modification was done quite earlier than the time in which it is put in the feed,
the mtime is older than the feed version and the plugin is no uploaded/updated in the redis cache.
This produced that some metadata or preferences is outdated for vts in this situation.
This patch fix this, and now the upload is always done for all vts to ensure that the cache is up-to-date.