Change: Don't wrap-up the scan.
Jira: SC-378
Jira: SC-34

Currently a wrap-up is performed for stopped scan, which launch plugins in the ACT_END category, before finishing.
This has more disadvantage than advantages. If a scan is stopped, it is supposed that no more info is necessary.
Also, in case of resuming the scan, all results of non-finished hosts will be removed.

This is causing to leave some results or keys in the host's kb, after the kbs were released, which could produce to mix results for the next started scan.

For testing:

Run a F&F scan, wait some minutes and stop it. Check in redis if there is some KB, with info, but it is not set as taken in the db namespace 0.