Change: when detecting a inconsistency kill parent and ourselves
When detecting a inconsistency we should not start an attack and stop
immediately.

This commit will kill the running process as well as the parent.

This is done because currently we don't have a proper communication
channel to all our ancestors.

This should be changed when introducing IPC to kill the oldest member of
our line to also kill all inconsistent processes instead of letting them
figure it our and end themselves.