Add: scan_id handling within misc, kb_check_push_str
To avoid using static char * within openvas.c and make it more
convenient to use scan_id handling is handled within scan_id.c by the
methods:

- set_scan_id - to set the scan_id
- get_scan_id - to get it

If the scan_id is not yet set then get_scan_id returns NULL.

To make it easier to verify the kb while trying to write there the
functions:

- kb_check_push_str (kb_t, const char *, const char *, size_t);
- kb_check_set_str (kb_t, const char *, const char *, size_t);
- kb_check_add_str_unique (kb_t, const char *, const char *, size_t, int);
- kb_check_set_int (kb_t, const char *, int);
- kb_check_add_int (kb_t, const char *, int);
- kb_check_add_int_unique (kb_t, const char *, int);

The call `check_kb_inconsistency` and if there is a global scan_id and
the db contains a scan_id then it verifies if they're the same.

If not those methods will not call the underlying `kb_item_` methods.

If there is no no scan_id in the kb it will print a debug message that
this call is maybe misplaced.
But it will store it within the given kb in those cases. This is done
intentionally to may figure cases in that `kb_item_push_str` is more
suitable.