Fix: PLugin process keeps scanning after manually stopping task
SIGUSR1 signal is sent during scan stop to all host processes, it will
call pluginlaunch_stop ();
There is a very special situation, when the last plugin of the last host
in the scan task is running, we wait for it to finish and it is not possible anymore to
check if the scan is stopped or not.

This patch ensure that the pluginlaunch_stop() will be called also in this special case,
sending the SIGTERM signal to the plugin process group, terminating the plugins child as well
in case it has forked.