Add: inter process communication
To be able to talk to send data downstream to children (and their
children) and upstream (to the parent and main process) IPC is
introduced.

This first and simplest use case is to append found vhosts via
add_hostname.

The usage of ipc is simplified by storing all children of a process in a
global variable within processes. The handling, creation, freeing etc is
done via specialized functions of processes.

There are three header files:
- ipc_pipe.h contains the definitions for pipe specific implementation
- ipc.h contains the definitions for usage
- ipc_openvas.h contains the definitions specific for openvas

This is done to make it easier to either extend IPC or replace specific
solutions without having to change the clients too much.