Add: Base documentation Tools
This PR adds tools, which allows to generate a HTML Page from the complete documentation and man pages of NASL library functions and programs within the OpenVAS project

The man-pages for the nasl library can be generated with './man.sh' or 'make nasl-man' (after build).

A HTML documentation can be generated with './html.sh' or 'make manual' (after build). It contains also the man pages as well as the redis documentation. In addition it contains a few discriptions of NASL built-in functions.