Handle User-Agent sent to the parent process via IPC (#1192)
What:
Via IPC, the first time that the User-Agent is built, can be sent to the parent process and stored. Later, the plugin processes inherits it and uses it.

Jira: SC-665

Why:
Currently, the user agent must be taken from a script preference, and in case this is not set, it will get from the vendor. As fall back it uses the nasl library version. This will be done thousand times during a F&F scan, since it is done each time a nasl script calls an nasl_http_* function.

How:
Run a F&F scan. Set vendor version in the openvas.conf or a User-Agent in the plugin preferences of global_settings.nasl.

* Add user functions to set/get the User-Agent

* Update cmake with the new files and add a new definition for OPENVAS_MISC_VERSION

* Uses the new function user_agent_get() instead of build it on the fly

* Add function to handle the user agent via IPC

* Handle the case in which a User-Agent is received for the parent process via IPC

* Send the User-Agent to the parent proces via IPC

* Send the ipc context, so the child can sent the info to the parent via ipc

* Fix format

* Add type definitions for ipc_hostname and ipc_user_agent structures

* Use union instead of void* inside the ipc_data struct

* Limit the access to the structures. Get ipc_data_t internal data only via ipc functions

* Initialize parser to fix warning ''parser' may be used uninitialized in this function'

* Add IP_DT_ERROR data type and handle the error.

Also, uses the function to get the type instead of accessing the data in the struct directly

* Improce code quality. Check for null and data type before referencing.

* Improve/Add docstrings

(cherry picked from commit c7ecbd0ff648ca32959ed4c850dd586c0d128673)