Fix: Bug reported by the clang static analyzer.
This issue was hard to fix, and still not sure if it is the right solution.
Please do a deep review and try running the scan analyzer again without this patch. See the report check if it is the right solution,
or I am introducing a new bug.

Description: Assigned value is garbage or undefined
File: /home/jnicola/dev/openvas/nasl/charcnv.c
Line: 625