Fix: use main_kb to verify scan_id for inconsistencies
This commit fixes the issue that the inconsistencies were not
necessarily checked against the shared database between ospd and
openvas.

Since ospd sets the scanner id into redis before calling openvas it is
necessary to use the main_kb to verify the locally set scanid with the
one in the database.

This is done by globally storing the main_kb via plugutils `set_main_kb`
and getting it via `get_main_kb`.