Change: switch build_token to macro, create macro to build double_tokens
To make the Iterator implementation for Tokenizer easier to read a new
macro double_token is created to simplify the switch cases.

Instead of having to define each case like:

```
'+' => {
  let next = $cursor.peek(0);
  match next {
    '+' => {
       self.cursor.advance();
       single_token!(PlusPlus, start, self.cursor.len_consumed())
    },
    '=' => {
       self.cursor.advance();
       single_token!(PlusEqual, start, self.cursor.len_consumed())
    },
    _ => single_token!(Plus, start, self.cursor.len_consumed()),
  }
}
```
 it is simplified to
```
'+' => double_token!(self.cursor, start, Plus, '+', PlusPlus, '=', PlusEqual),
```
which improves the readability for those standard cases.

For this purpose the lambda build_token was redefined as a macro
single_token! so it is useable within the token crate.