Fix bug on AssignReturn, simplify Block
Fixes a bug on AssignReturn that used the token of the ++/-- operator
instead of the variable within the operation.

Removed Block struct in favor of a Statement and Box depth.

Since we need to store the block variables and function definitions
elsewhere anyway we don't need references and values but an Identifier.

Currently this Identifier is defined as (u8, u8) while the first value
represents the index of the parent while the second one is the own
index.