Change unhandled_token, abort condition and move DoublePoint to infix
Due to the confusing handling of self.end_category as well as
unhandled_token within the Lexer it returns a boolean if a statement is
finished or not.

The reason why it needs a boolean instead of a failure result is that
operations can be nested and therefore sometimes require the possibility
to return unfinished statements for further processing.

A aftermath of that is that end_category is not longer needed and the
reset of unhandled_token is also not required anymore.

The DoublePoint handling moved from a Prefix to a Infix since it is a
lot easier to handle `name:value` as a infix.