Refactor: make nasl-cli somewhat useful, started to use ErrorKind enum
To use nasl-cli instead of a test case it can recursively go through
dirs and parse nasl and inc files in that dir. It than prints a summary
if error got found.

Additionally errors within nasl-syntax now contain a enum ErrorKind
instead of a reason string to make it easier to pattern match against
it.