Refactor: move string resolver into StringCategory
Creates a trait called Resolver<T> with a resolve methods and implements
it for StringCategory instead of a method in the Interpreter.

This makes it easier to resolve it since it is just relevant for
Category::String that contains a StringCategory.

Implements From<Token> for Range instead of a range method to be more
standardized.