Change: context handling
Instead of having a trait to deal with context it is now implemented as
a struct and is stored within a register.

The idea here is that on each new context the register is called to a
create a context and after a context is finished it must be dropped.