Redfine function creation in description
Instead of calling two macros it is defined as one:
`make_storage_function`
with a small DSL to create specific storage functions for the
description block.

The small DSL is defined as

```
make_storage_function! {
  script_timeout => [timeout :1],
  script_category => [category :1],
  script_name => [name :1],
  script_version => [version :1],
  script_copyright => [copyright :1],
  script_family => [family :1],
  script_oid => [oid :1],
  script_dependencies => [dependencies :0],
  script_mandatory_keys => [mandatory_keys: 0],
  script_require_ports => [required_ports: 2],
  script_tag => (name: value)
}
```

The first parameter is the function name followed by an `=>` the
optional positional argument handling is defined within `[]` the first
parameter within this block is the key name for the positional arguments
followed by an `:` to define the expected positional arguments or 0 if
there is no limit.

The named parameter are defined within an `()` the first argument is the
name of the key value followed by an `:` and the name of the value
argument.