Introduce sink library
Instead of using a complete struct (like in `nvtcache` crate) or just
use str, str like in the `nasl-interpreter` crate it works with Fields.

Currently there is just NVT supported so a NVT get split up into:

- Oid(String),
- FileName(String),
- Name(String),
- Tag(TagKey, String),
- Dependencies(Vec<String>),
- RequiredKeys(Vec<String>),
- MandatoryKeys(Vec<String>),
- ExcludedKeys(Vec<String>),
- RequiredPorts(Vec<String>),
- RequiredUdpPorts(Vec<String>),
- Preference(NvtPreference),
- Reference(NvtRef),
- Category(ACT),
- Family(String),
- NoOp - those fields are getting ignored.

Those will then be send by the interpreter to the storage.

Unfortunately the current redis implementation needs a complete struct
to work with therefore a on_exit method is declared that is also called
by the interpreter.

In the redis implementation the on_exit method iterates through the
stored fields and creates the necessary struct for the actual storing
process.

The name is sink and not storage because theoretically we could also use
it to distribute messages into a broker or similar systtem and are not
limited to a storing implementation.