Refactor: move the context handling to the register
Instead of bothering the caller to deal with last_mut there is a new
method to add parameter called `add_local`.

In NASL the default are global parameter, hence add_local is only needed
on declare local.

However to make it easier to change later on it is moved to Register.