Change use reference in interpreter::resolve
Since the statement within resolve of interpreter does not mutate the
statement but reads it to generate Values it is changed to a reference
instead.

This makes it clear that Interpreter does not own the Statement and that
it is reading it making it easier to handle and clearer in the future.