Add `feed-verifier`
`feed-verifier` is a specialized program to verify the output of
`openvas -u` with the `nasl-cli` implementation by:
- flushing redis
- executing `openvas -u`
- get the results
- flushing redis
- executing `nasl-cli feed update`
- get the results
- verify results of `openvas -u` with results of `nasl-cli feed updaye`

It does not require any arguments since it assumes that `nasl-cli` is in
the same dir as the own executable and that `openvas` is in `PATH`.

The redis connection parameter as well as plugin path it gets from `openvas -s`.