Adds test to verify nasl-cli feed update with openvas -u
Adds a tests that verify `openvas -u` and `nasl-cli feed update` behave the
same by executing `feed-verifier`.

Since we support `debian:stable` it additionally checks that `nasl-cli` is
able to compile with libc 2.31 by utilizing ubuntu-20.4 instead of the latest
ubuntu version.

It also removed the dependency on dns-lookup in favor of
std::net::ToSocketAddrs.