Change load feed over sha256sums
Changes nasl-cli to load the files via the sha256sums.

This improves the behaviour in two ways:
1. it is faster because the filesystem does not need to be recursively
   scanned for .nasl files
2. It only runs nasl scripts in description mode that have a valid
   sha256 sum.

This is done by creating a feed library that includes Update
functionality. The Update is written as an iterator to allow handling
each case individually in the program in a standard fashion.

The underlying FSPluginLoader is changed to allow loading files by line,
this allows streaming through the hashsum file instead of having to load
each filename upfront.

The hashing algorithm can be chosen when there multiple implementation
by calling `HashSumNameLoader::sha256` currently only sha256 is
supported.

To use the Update command you can execute:

```
    let loader = FSPluginLoader::new(path);
    let verifier = feed::HashSumNameLoader::sha256(&loader)?;
    let updater = feed::Update::init("1", 5, loader.clone(), storage, verifier);
    for s in updater {
      println!("updated {s}");
    }
```