Add libpcap to the cross docker image
In preparation of using pcap the cross build is extended to add it.

Currently the aarch64 image is building pcap manually because currently
the stable image is based on xenial which doesn't provide a pre-build
version of it for aarch64. This will change with the next stable version
as edge is based on 20.4 which has a prebuild version of it.