Add PerNvtSink generic implementation
For the cases that require an disptach when a description run per file
is finished a generic implementation for caching is implement.

Additional changes are:

Refactor use Default implementation for Nvt.

Changed description run to send filename on script_oid based on the key
instead of enforcing the user to know this implementation detail.

This makes it easier to create new sinks as it is handled within the
description run.

Introduces `TagValue` to handle special tags like:

- TagKey::CreationDate
- TagKey::LastModification
- TagKey::SeverityDate

that will be parsed to Unix timestamp.

This moves the control to handle the TagKey to sink instead of the
implementation and has the goal to make it less error prone to implement
a new sink implementation.

Removes unused retrieve method from `sink::Sink`.

Changes redis-sink to use new PerNvtSink method.