Add json-sink
Adds a specialized sink to convert fields to json.

Currently it supports NVT -> Json as either an array run or as a single
json.

To transform a feed nasl-cli is extended with `feed transform` that
transforms a feed to a json array in stdout. The intended purpose is
that it is run with piping in mind
`nasl-cli feed transform > feed.json`.

nasl-cli transform requires a sha256sums file.

Changes serde to be an optional requirement within the sink.

Add an From<io::Error> implementation for SinkError to handle standard
cases in an easier fashion.

Changes the redis-sink RedisCache to NvtDispatcher to have a similar
contract as a json-sink.

Change nasl-cli to ignore BrokenPipe errors instead panic to make | head
piping easier.

Remove unused text within NvtReference.