Add KB dispatcher enum, refactor types to be defined in sink
To allow dispatching KB items a new enum KB containing a key, value and
an optional expire is add.

To handle the various types within an KB item the types are refactored
to be set in sink/types.rs. Currently NASlValue does translate between
those but it may makes sense to reuse them within NASlValue instead.