Change key to be generic instead of string
In future runs we probably need a key based of: hostname and oid
therefore we should be more generic on what kind of keys we allow to
make it easier to switch the key handling later on.

The only requirement which needs to be fulfilled is that a key must
implement `AsRef<str>`.