Use FunctionErrorKind instead of FunctionError within builtin functions
To make it easier to handle FunctionError while writing builtin function
the NaslFunction is changed to return a FunctionErrorKind instead.

With this change the function name will be set on call when an error
occurs rather than enforcing the developer to copy and paste the
function name while creating the error.