Refactor: remove `Nothing` mode, create log method
Instead of having to individually implement the debug, info, warn, and error
methods, these methods are already implemented within the trait.

The only method that needs to be implemented is the log method.

Furthermore, the "Nothing" mode has been removed, and the display method now
uses the info method. This decision was made because, during a multi-scan run,
information that is at a higher level than info (such as warning or error) is
not only useless but also difficult to configure a log aggregator
without a widely known log level.