CI: Implement a new release process for multi-project repository
This commit updates our CI release process to accommodate the switch from a
single project repository to a multi-project one. The new release process
includes the following changes:

1. We now base releases on git tags instead of individual projects.
1. We use major, minor, and patch releases instead of just patch releases.
1. Releases are made from the main branch instead of stable or unstable branches.
1. For major releases, we create a version branch to handle hotfixes.
1. Each project within the repository has the same version.
1. We've added nasl-cli as a release artifact for linux-x86_64 and aarch64.

To initiate the release process, you can either add one of the following labels
to a pull request:
- `major_release`,
- `minor_release`,
- `patch_release`.

Once the PR is merged, the release process will be triggered.

Alternatively, you can initiate the release via the GitHub action page.

Note that we are currently working on creating a docker image for releases,
which will require new docker tag layout and compose changes.